+++ CMakeLists.txt
@@ -370,6 +370,10 @@ set(API_PLUGIN_FILES
         web/api/web_api_v1.h
         web/api/badges/web_buffer_svg.c
         web/api/badges/web_buffer_svg.h
+        web/api/exporters/allmetrics.c
+        web/api/exporters/allmetrics.h
+        web/api/exporters/shell/allmetrics_shell.c
+        web/api/exporters/shell/allmetrics_shell.h
         )
 
 set(STREAMING_PLUGIN_FILES
