+++ web/api/badges/web_buffer_svg.h
@@ -8,4 +8,6 @@
 extern void buffer_svg(BUFFER *wb, const char *label, calculated_number value, const char *units, const char *label_color, const char *value_color, int precision, int scale, uint32_t options);
 extern char *format_value_and_unit(char *value_string, size_t value_string_len, calculated_number value, const char *units, int precision);
 
+extern int web_client_api_request_v1_badge(RRDHOST *host, struct web_client *w, char *url);
+
 #endif /* NETDATA_WEB_BUFFER_SVG_H */
