+++ web/api/exporters/prometheus/README.md
@@ -1,0 +1,3 @@
+# prometheus exporter
+
+The prometheus exporter for netdata is located at the [backends section for prometheus](../../../../backends/prometheus).
