+++ web/api/rrd2json.h
@@ -4,6 +4,7 @@
 #define NETDATA_RRD2JSON_H 1
 
 #include "web_api_v1.h"
+#include "exporters/allmetrics.h"
 
 #define HOSTNAME_MAX 1024
 
@@ -35,16 +36,6 @@
 #define DATASOURCE_FORMAT_SSV_COMMA "ssvcomma"
 #define DATASOURCE_FORMAT_CSV_JSON_ARRAY "csvjsonarray"
 
-#define ALLMETRICS_FORMAT_SHELL                 "shell"
-#define ALLMETRICS_FORMAT_PROMETHEUS            "prometheus"
-#define ALLMETRICS_FORMAT_PROMETHEUS_ALL_HOSTS  "prometheus_all_hosts"
-#define ALLMETRICS_FORMAT_JSON                  "json"
-
-#define ALLMETRICS_SHELL                        1
-#define ALLMETRICS_PROMETHEUS                   2
-#define ALLMETRICS_JSON                         3
-#define ALLMETRICS_PROMETHEUS_ALL_HOSTS         4
-
 #define GROUP_UNDEFINED         0
 #define GROUP_AVERAGE           1
 #define GROUP_MIN               2
@@ -71,9 +62,6 @@
 
 extern void rrd_stats_api_v1_chart(RRDSET *st, BUFFER *wb);
 extern void rrd_stats_api_v1_charts(RRDHOST *host, BUFFER *wb);
-
-extern void rrd_stats_api_v1_charts_allmetrics_json(RRDHOST *host, BUFFER *wb);
-extern void rrd_stats_api_v1_charts_allmetrics_shell(RRDHOST *host, BUFFER *wb);
 
 extern int rrdset2anything_api_v1(
           RRDSET *st
