+++ web/api/web_api_v1.h
@@ -17,9 +17,7 @@ extern int web_client_api_request_v1_alarm_log(RRDHOST *host, struct web_client 
 extern int web_client_api_request_single_chart(RRDHOST *host, struct web_client *w, char *url, void callback(RRDSET *st, BUFFER *buf));
 extern int web_client_api_request_v1_alarm_variables(RRDHOST *host, struct web_client *w, char *url);
 extern int web_client_api_request_v1_charts(RRDHOST *host, struct web_client *w, char *url);
-extern int web_client_api_request_v1_allmetrics(RRDHOST *host, struct web_client *w, char *url);
 extern int web_client_api_request_v1_chart(RRDHOST *host, struct web_client *w, char *url);
-extern int web_client_api_request_v1_badge(RRDHOST *host, struct web_client *w, char *url);
 extern int web_client_api_request_v1_data(RRDHOST *host, struct web_client *w, char *url);
 extern int web_client_api_request_v1_registry(RRDHOST *host, struct web_client *w, char *url);
 extern int web_client_api_request_v1(RRDHOST *host, struct web_client *w, char *url);
